Move the files 'Memory.c' and 'Memory.h' into the folder 'payload/sp'
