+++ configure.py
@@ -200,7 +200,6 @@ code_in_files = {
         os.path.join('payload', 'game', 'system', 'HomeButton.S'),
         os.path.join('payload', 'game', 'system', 'InputManager.S'),
         os.path.join('payload', 'game', 'system', 'InputManager.c'),
-        os.path.join('payload', 'game', 'system', 'Memory.c'),
         os.path.join('payload', 'game', 'system', 'Mii.S'),
         os.path.join('payload', 'game', 'system', 'MultiDvdArchive.S'),
         os.path.join('payload', 'game', 'system', 'MultiDvdArchive.c'),
@@ -305,6 +304,7 @@ code_in_files = {
         os.path.join('payload', 'sp', 'IOSKeyboard.c'),
         os.path.join('payload', 'sp', 'Keyboard.c'),
         os.path.join('payload', 'sp', 'LogFile.c'),
+        os.path.join('payload', 'sp', 'Memory.c'),
         os.path.join('payload', 'sp', 'Net.c'),
         os.path.join('payload', 'sp', 'NetStorage.c'),
         os.path.join('payload', 'sp', 'NetStorageClient.c'),
